Scaffold REST endpoint and consolidate code for reuse
